+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Improvements for memories
   - Added new display modes: Fader, flash button or toggle button
   - Added mirroring mode for gradients
+  - Added state and gradient ordering by map coordinates
 - Improvements for dynamic pages
   - Filter out already selected entities when selecting more
   - Selected entities can be reordered via drag & drop